+++ pybuildkite/metrics.py
@@ -1,6 +1,5 @@
 from posixpath import join as urljoin
-from pybuildkite.client import Client
-from pybuildkite.constants import APIVersion
+from pybuildkite.client import AgentClient as Client
 
 
 class Metrics(Client):
@@ -8,7 +7,7 @@ class Metrics(Client):
     Metrics operations for the Agent API
     """
 
-    def __init__(self, client: Client, base_url: str):
+    def __init__(self, client: Client, base_url: str) -> None:
         """
         Construct the class
 
@@ -16,16 +15,11 @@ class Metrics(Client):
         :param base_url: Base Url
         """
         self.client = client
-        self.client._api_version = APIVersion.V3
         self.path = urljoin(base_url, "metrics")
 
-    def get_metrics(
-        self,
-    ):
+    def get_metrics(self):
         """
         Returns a paginated list of an organization’s agents.
 
         """
-        return self.client.get(
-            self.path,
-        )
+        return self.client.get(self.path)
